# TraceQL Metrics: Fix incorrect optimization for some queries